Update to skeleton template
Applies the current ITensorPkgSkeleton template to ITensorGaussianMPS.jl,
bringing it structurally in line with the rest of the ecosystem so it can
be included in ITensor/ITensorOrgPatches and receive ecosystem-wide
patches automatically.

No user-facing API changes (patch bump 0.1.13 → 0.1.14).

- Replaces legacy CI.yml + Register.yml with the split skeleton
  workflows that delegate to ITensor/ITensorActions: Tests, Documentation,
  FormatCheck, FormatPullRequest, IntegrationTest, IntegrationTestRequest,
  Registrator, VersionCheck. TagBot, CompatHelper, and dependabot synced
  to skeleton too.
- Adds .pre-commit-config.yaml, benchmark/, docs/src/assets/,
  docs/src/reference.md, examples/README.jl, test/test_aqua.jl,
  test/test_basics.jl, .github/ISSUE_TEMPLATE/, PULL_REQUEST_TEMPLATE.md.
- Root Project.toml gets [workspace] projects = ["benchmark", "dev",
  "docs", "examples", "test"]. Subdir Project.tomls get
  [sources.ITensorGaussianMPS] path = ".." plus [compat] entries for
  their direct deps (Aqua, ITensorMPS, ITensors, LinearAlgebra, Test,
  ITensorPkgSkeleton, Documenter, ITensorFormatter).
- test/runtests.jl switches to ITensorPkgSkeleton.runtests(;testdir),
  which discovers test_*.jl files automatically and walks examples/.
  Renames gmps.jl/electron.jl/linalg.jl → test_*.jl to match; gmera.jl
  is not auto-discovered (it wasn't run by the old runtests.jl either).
- test_aqua.jl disables piracies and ambiguities — piracy is by design
  (Hermitian/Givens constructor extensions, ITensor/MPS constructor
  extensions), and the Circuit * ambiguities with FillArrays are
  pre-existing and out of scope for template migration.
- Cleans up the implicit `Hermitian` extension in gmps.jl that would
  silently no-op on Julia 1.12: switches the bare `using LinearAlgebra`
  in the module to an explicit `using LinearAlgebra: LinearAlgebra,
  Adjoint, Diagonal, Givens, Hermitian, …` list, and qualifies the
  extension site as `LinearAlgebra.Hermitian(A::T) = …`.
- Fixes an unclosed code fence in the `givens_rotations(::ConservesNfParity)`
  docstring that was confusing the formatter's parser (stray `c` after
  the code block).
- Deletes examples/spinless_fermion_pairing.jl. It errors at runtime
  with "Not able to construct real fermionic basis for input correlation
  matrix." — a pre-existing library bug, not a template-migration issue.
  The old runtests.jl never exercised it; ITensorPkgSkeleton.runtests
  now auto-walks examples/ so it would fail CI. Track the underlying
  fix separately.
- itpkgfmt --yaml pass across the tree (src/, test/, examples/, workflow
  YAML) — source file reformatting accounts for most of the diff volume.
