+++ .github/ISSUE_TEMPLATE/BUG_REPORT.md
@@ -1,0 +1,62 @@
+---
+name: ITensorGaussianMPS.jl bug report
+about: Create a bug report to help us improve ITensorGaussianMPS.jl
+title: "[BUG] YOUR SHORT DESCRIPTION OF THE BUG HERE"
+labels: ["bug"]
+assignees: ''
+
+---
+
+**Description of bug**
+
+Please give a brief description of the bug or unexpected behavior here.
+
+**Minimal code demonstrating the bug or unexpected behavior**
+
+If applicable, provide a minimal code that can be run to demonstrate the bug or unexpected behavior.
+
+If you are unable to construct a minimal code that demonstrates the bug or unexpected behavior, provide detailed steps for how to reproduce the behavior you are seeing.
+
+<details><summary>Minimal runnable code</summary><p>
+
+```julia
+[YOUR MINIMAL RUNNABLE CODE HERE]
+```
+
+</p></details>
+
+
+**Expected output or behavior**
+
+Describe what you expected to happen.
+
+If you provided a minimal code that can be run to demonstrate the bug or unexpected behavior, describe what you expected the output would be.
+
+
+**Actual output or behavior**
+
+Describe what actually happened.
+
+If you provided a minimal code that demonstrates the bug or unexpected behavior, provide the output you get from that code. If the code leads to an error or warning, include the full error or warning below.
+
+<details><summary>Output of minimal runnable code</summary><p>
+
+```julia
+[OUTPUT OF YOUR MINIMAL RUNNABLE CODE HERE]
+```
+
+</p></details>
+
+
+**Version information**
+
+ - Output from `versioninfo()`:
+```julia
+julia> versioninfo()
+[YOUR OUTPUT HERE]
+```
+ - Output from `using Pkg; Pkg.status("ITensorGaussianMPS")`:
+```julia
+julia> using Pkg; Pkg.status("ITensorGaussianMPS")
+[YOUR OUTPUT HERE]
+```
