+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,42 @@
+# Description
+
+Please include a summary of the change and which issue is fixed (if applicable). Please also include relevant motivation and context. List any dependencies that are required for this change.
+
+Fixes #(issue)
+
+If practical and applicable, please include a minimal demonstration of the previous behavior and new behavior below.
+
+<details><summary>Minimal demonstration of previous behavior</summary><p>
+
+```julia
+[YOUR MINIMAL DEMONSTRATION OF PREVIOUS BEHAVIOR]
+```
+
+</p></details>
+
+<details><summary>Minimal demonstration of new behavior</summary><p>
+
+```julia
+[YOUR MINIMAL DEMONSTRATION OF NEW BEHAVIOR]
+```
+
+</p></details>
+
+# How Has This Been Tested?
+
+Please add tests that verify your changes to a file in the `test` directory.
+
+Please give a summary of the tests that you added to verify your changes.
+
+- [ ] Test A
+- [ ] Test B
+
+# Checklist:
+
+- [ ] My code follows the style guidelines of this project. Please run the [ITensorFormatter](https://github.com/ITensor/ITensorFormatter.jl) in the base directory of the repository (`~/.julia/dev/ITensorGaussianMPS`) to format your code according to our style guidelines.
+- [ ] I have performed a self-review of my own code.
+- [ ] I have commented my code, particularly in hard-to-understand areas.
+- [ ] I have added tests that verify the behavior of the changes I made.
+- [ ] I have made corresponding changes to the documentation.
+- [ ] My changes generate no new warnings.
+- [ ] Any dependent changes have been merged and published in downstream modules.
