+++ .github/workflows/FormatCheck.yml
@@ -1,0 +1,16 @@
+name: "Format Check"
+on:
+  pull_request_target:
+    types:
+      - "opened"
+      - "synchronize"
+      - "reopened"
+      - "ready_for_review"
+permissions:
+  contents: "read"
+  actions: "write"
+  pull-requests: "write"
+jobs:
+  format-check:
+    name: "Format Check"
+    uses: "ITensor/ITensorActions/.github/workflows/FormatCheck.yml@main"
