+++ .github/workflows/FormatPullRequest.yml
@@ -1,0 +1,16 @@
+name: "Format Pull Request"
+on:
+  schedule:
+    - cron: "0 0 * * *"
+  workflow_dispatch: ~
+  issue_comment:
+    types:
+      - "created"
+permissions:
+  contents: "write"
+  pull-requests: "write"
+jobs:
+  format-pull-request:
+    name: "Format Pull Request"
+    uses: "ITensor/ITensorActions/.github/workflows/FormatPullRequest.yml@main"
+    secrets: "inherit"
