+++ .github/workflows/IntegrationTest.yml
@@ -1,0 +1,37 @@
+name: "IntegrationTest"
+on:
+  push:
+    branches:
+      - "main"
+    tags: "*"
+  pull_request_target:
+    types:
+      - "opened"
+      - "synchronize"
+      - "reopened"
+      - "ready_for_review"
+      - "converted_to_draft"
+jobs:
+  integration-test:
+    name: "IntegrationTest"
+    strategy:
+      fail-fast: false
+      matrix:
+        pkg:
+          - "__none__"
+    uses: "ITensor/ITensorActions/.github/workflows/IntegrationTest.yml@main"
+    secrets: "inherit"
+    with:
+      localregistry: "https://github.com/ITensor/ITensorRegistry.git"
+      pkg: "${{ matrix.pkg }}"
+  integration-gate:
+    name: "IntegrationTest"
+    needs: "integration-test"
+    if: "${{ always() && needs.integration-test.result != 'skipped' }}"
+    runs-on: "ubuntu-latest"
+    steps:
+      - name: "Fail if any downstream integration test failed"
+        run: |
+          echo "integration-test.result = ${{ needs.integration-test.result }}"
+          test "${{ needs.integration-test.result }}" = "success"
+
