+++ .github/workflows/IntegrationTestRequest.yml
@@ -1,0 +1,14 @@
+name: "Integration Test Request"
+on:
+  issue_comment:
+    types:
+      - "created"
+jobs:
+  integrationrequest:
+    if: |
+      github.event.issue.pull_request &&
+      contains(fromJSON('["OWNER", "COLLABORATOR", "MEMBER"]'), github.event.comment.author_association)
+      
+    uses: "ITensor/ITensorActions/.github/workflows/IntegrationTestRequest.yml@main"
+    with:
+      localregistry: "https://github.com/ITensor/ITensorRegistry.git"
