+++ .github/workflows/Registrator.yml
@@ -1,0 +1,22 @@
+name: "Register Package"
+on:
+  workflow_dispatch: ~
+  push:
+    branches:
+      - "master"
+      - "main"
+    paths:
+      - "Project.toml"
+  issue_comment:
+    types:
+      - "created"
+permissions:
+  contents: "write"
+  pull-requests: "write"
+  issues: "write"
+jobs:
+  Register:
+    uses: "ITensor/ITensorActions/.github/workflows/Registrator.yml@main"
+    with:
+      localregistry: "ITensor/ITensorRegistry"
+    secrets: "inherit"
