+++ .github/workflows/TagBot.yml
@@ -1,31 +1,11 @@
-name: TagBot
+name: "TagBot"
 on:
   issue_comment:
     types:
-      - created
-  workflow_dispatch:
-    inputs:
-      lookback:
-        default: "3"
-permissions:
-  actions: read
-  checks: read
-  contents: write
-  deployments: read
-  issues: read
-  discussions: read
-  packages: read
-  pages: read
-  pull-requests: read
-  repository-projects: read
-  security-events: read
-  statuses: read
+      - "created"
+  workflow_dispatch: ~
 jobs:
   TagBot:
-    if: github.event_name == 'workflow_dispatch' || github.actor == 'JuliaTagBot'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: JuliaRegistries/TagBot@v1
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-          ssh: ${{ secrets.DOCUMENTER_KEY }}
+    if: "github.event_name == 'workflow_dispatch' || github.actor == 'JuliaTagBot'"
+    uses: "ITensor/ITensorActions/.github/workflows/TagBot.yml@main"
+    secrets: "inherit"
