+++ .gitignore
@@ -1,2 +1,15 @@
-Manifest.toml
-.*.swp
+*.cov
+*.mem
+*.o
+*.swp
+.DS_Store
+.benchmarkci
+.tmp
+.vscode/
+.worktrees/
+LocalPreferences.toml
+Manifest*.toml
+benchmark/*.json
+dev/
+docs/build/
+docs/src/index.md
