+++ .pre-commit-config.yaml
@@ -1,0 +1,17 @@
+ci:
+  skip:
+    - "itensor-formatter"
+repos:
+  - repo: "https://github.com/pre-commit/pre-commit-hooks"
+    rev: "v6.0.0"
+    hooks:
+      - id: "check-merge-conflict"
+      - id: "check-toml"
+      - id: "check-yaml"
+      - id: "end-of-file-fixer"
+        exclude_types:
+          - "markdown"
+  - repo: "https://github.com/ITensor/itensorformatter-pre-commit"
+    rev: "v1.0.0"
+    hooks:
+      - id: "itensor-formatter"
