+++ docs/src/reference.md
@@ -1,0 +1,5 @@
+# Reference
+
+```@autodocs
+Modules = [ITensorGaussianMPS]
+```
