+++ examples/Project.toml
@@ -1,8 +1,8 @@
 [deps]
 ITensorGaussianMPS = "2be41995-7c9f-4653-b682-bfa4e7cebb93"
-ITensorMPS = "0d1a4710-d33b-49a5-8f18-73bdf49b47e2"
-ITensors = "9136182c-28ba-11e9-034c-db9fb085ebd5"
+
+[sources.ITensorGaussianMPS]
+path = ".."
 
 [compat]
-ITensorMPS = "0.2.2"
-ITensors = "0.6.7"
+ITensorGaussianMPS = "0.1"
