+++ examples/hubbard_1d_spin_conservation.jl
@@ -19,7 +19,7 @@ _maxlinkdim = 200
 @show _maxlinkdim
 
 # DMRG cutoff
-_cutoff = 1e-8
+_cutoff = 1.0e-8
 
 # Hopping
 t = 1.0
@@ -32,15 +32,15 @@ U = 1.0
 # Make the free fermion Hamiltonian for the up spins
 os_up = OpSum()
 for n in 1:(N - 1)
-  os_up .+= -t, "Cdagup", n, "Cup", n + 1
-  os_up .+= -t, "Cdagup", n + 1, "Cup", n
+    os_up .+= -t, "Cdagup", n, "Cup", n + 1
+    os_up .+= -t, "Cdagup", n + 1, "Cup", n
 end
 
 # Make the free fermion Hamiltonian for the down spins
 os_dn = OpSum()
 for n in 1:(N - 1)
-  os_dn .+= -t, "Cdagdn", n, "Cdn", n + 1
-  os_dn .+= -t, "Cdagdn", n + 1, "Cdn", n
+    os_dn .+= -t, "Cdagdn", n, "Cdn", n + 1
+    os_dn .+= -t, "Cdagdn", n + 1, "Cdn", n
 end
 
 # Hopping Hamiltonians for the up and down spins
@@ -52,20 +52,20 @@ h_dn = hopping_hamiltonian(os_dn)
 Φ_dn = slater_determinant_matrix(h_dn, Nf_dn)
 
 # Create an MPS from the slater determinants.
-s = siteinds("Electron", N; conserve_qns=true)
+s = siteinds("Electron", N; conserve_qns = true)
 println("Making free fermion starting MPS")
 @time ψ0 = slater_determinant_to_mps(
-  s, Φ_up, Φ_dn; eigval_cutoff=1e-4, cutoff=_cutoff, maxdim=_maxlinkdim
+    s, Φ_up, Φ_dn; eigval_cutoff = 1.0e-4, cutoff = _cutoff, maxdim = _maxlinkdim
 )
 @show maxlinkdim(ψ0)
 
 # The total non-interacting part of the Hamiltonian
 os_noninteracting = OpSum()
 for n in 1:(N - 1)
-  os_noninteracting .+= -t, "Cdagup", n, "Cup", n + 1
-  os_noninteracting .+= -t, "Cdagdn", n, "Cdn", n + 1
-  os_noninteracting .+= -t, "Cdagup", n + 1, "Cup", n
-  os_noninteracting .+= -t, "Cdagdn", n + 1, "Cdn", n
+    os_noninteracting .+= -t, "Cdagup", n, "Cup", n + 1
+    os_noninteracting .+= -t, "Cdagdn", n, "Cdn", n + 1
+    os_noninteracting .+= -t, "Cdagup", n + 1, "Cup", n
+    os_noninteracting .+= -t, "Cdagdn", n + 1, "Cdn", n
 end
 
 H_noninteracting = MPO(os_noninteracting, s)
@@ -75,13 +75,13 @@ H_noninteracting = MPO(os_noninteracting, s)
 # The total interacting Hamiltonian
 os_interacting = OpSum()
 for n in 1:(N - 1)
-  os_interacting .+= -t, "Cdagup", n, "Cup", n + 1
-  os_interacting .+= -t, "Cdagdn", n, "Cdn", n + 1
-  os_interacting .+= -t, "Cdagup", n + 1, "Cup", n
-  os_interacting .+= -t, "Cdagdn", n + 1, "Cdn", n
+    os_interacting .+= -t, "Cdagup", n, "Cup", n + 1
+    os_interacting .+= -t, "Cdagdn", n, "Cdn", n + 1
+    os_interacting .+= -t, "Cdagup", n + 1, "Cup", n
+    os_interacting .+= -t, "Cdagdn", n + 1, "Cdn", n
 end
 for n in 1:N
-  os_interacting .+= U, "Nupdn", n
+    os_interacting .+= U, "Nupdn", n
 end
 H = MPO(os_interacting, s)
 #@show norm(prod(H) - prod(H_noninteracting))
@@ -98,12 +98,12 @@ println("Free fermion starting state energy")
 @show inner(ψ0', H, ψ0)
 
 println("\nStart from random product state")
-er, ψ̃r = @time dmrg(H, ψr; nsweeps=10, maxdim=[10, 20, _maxlinkdim], cutoff=_cutoff)
+er, ψ̃r = @time dmrg(H, ψr; nsweeps = 10, maxdim = [10, 20, _maxlinkdim], cutoff = _cutoff)
 @show er
 @show flux(ψ̃r)
 
 println("\nStart from free fermion state")
-e0, ψ̃0 = @time dmrg(H, ψ0; nsweeps=5, maxdim=_maxlinkdim, cutoff=_cutoff)
+e0, ψ̃0 = @time dmrg(H, ψ0; nsweeps = 5, maxdim = _maxlinkdim, cutoff = _cutoff)
 @show e0
 @show flux(ψ̃0)
