+++ examples/hubbard_2d_spin_conservation.jl
@@ -21,7 +21,7 @@ _maxlinkdim = 1_000
 @show _maxlinkdim
 
 # DMRG cutoff
-_cutoff = 1e-5
+_cutoff = 1.0e-5
 
 # Hopping
 t = 1.0
@@ -31,20 +31,20 @@ U = 4.0
 
 @show t, U
 
-lattice = square_lattice(Nx, Ny; yperiodic=true)
+lattice = square_lattice(Nx, Ny; yperiodic = true)
 
 # Make the free fermion Hamiltonian for the up spins
 os_up = OpSum()
 for b in lattice
-  os_up .+= -t, "Cdagup", b.s1, "Cup", b.s2
-  os_up .+= -t, "Cdagup", b.s2, "Cup", b.s1
+    os_up .+= -t, "Cdagup", b.s1, "Cup", b.s2
+    os_up .+= -t, "Cdagup", b.s2, "Cup", b.s1
 end
 
 # Make the free fermion Hamiltonian for the down spins
 os_dn = OpSum()
 for b in lattice
-  os_dn .+= -t, "Cdagdn", b.s1, "Cdn", b.s2
-  os_dn .+= -t, "Cdagdn", b.s2, "Cdn", b.s1
+    os_dn .+= -t, "Cdagdn", b.s1, "Cdn", b.s2
+    os_dn .+= -t, "Cdagdn", b.s2, "Cdn", b.s1
 end
 
 # Hopping Hamiltonian with 2*N spinless fermions,
@@ -62,22 +62,22 @@ println()
 
 # Create an MPS from the slater determinant.
 # For now it only works without Sz conservation, this will be supported soon.
-s = siteinds("Electron", N; conserve_qns=true)
+s = siteinds("Electron", N; conserve_qns = true)
 println("Making free fermion starting MPS")
 @time ψ0 = slater_determinant_to_mps(
-  s, Φ_up, Φ_dn; eigval_cutoff=1e-4, cutoff=_cutoff, maxdim=_maxlinkdim
+    s, Φ_up, Φ_dn; eigval_cutoff = 1.0e-4, cutoff = _cutoff, maxdim = _maxlinkdim
 )
 @show maxlinkdim(ψ0)
 
 os = OpSum()
 for b in lattice
-  os .+= -t, "Cdagup", b.s1, "Cup", b.s2
-  os .+= -t, "Cdagdn", b.s1, "Cdn", b.s2
-  os .+= -t, "Cdagup", b.s2, "Cup", b.s1
-  os .+= -t, "Cdagdn", b.s2, "Cdn", b.s1
+    os .+= -t, "Cdagup", b.s1, "Cup", b.s2
+    os .+= -t, "Cdagdn", b.s1, "Cdn", b.s2
+    os .+= -t, "Cdagup", b.s2, "Cup", b.s1
+    os .+= -t, "Cdagdn", b.s2, "Cdn", b.s1
 end
 for n in 1:N
-  os .+= U, "Nupdn", n
+    os .+= U, "Nupdn", n
 end
 H = MPO(os, s)
 
@@ -94,14 +94,14 @@ println("\nFree fermion MPS starting state energy")
 
 println("\nStart from random product state")
 dmrg_kwargs = (;
-  nsweeps=10,
-  maxdim=[10, 20, 100, 200, _maxlinkdim],
-  cutoff=_cutoff,
-  noise=[1e-7, 1e-8, 1e-10, 0.0],
+    nsweeps = 10,
+    maxdim = [10, 20, 100, 200, _maxlinkdim],
+    cutoff = _cutoff,
+    noise = [1.0e-7, 1.0e-8, 1.0e-10, 0.0],
 )
 @time dmrg(H, ψr; dmrg_kwargs...)
 
 println("\nStart from free fermion state")
-@time dmrg(H, ψ0; nsweeps=10, maxdim=_maxlinkdim, cutoff=_cutoff)
+@time dmrg(H, ψ0; nsweeps = 10, maxdim = _maxlinkdim, cutoff = _cutoff)
 
 nothing
