+++ examples/mps_to_determinants.jl
@@ -17,7 +17,7 @@ _maxlinkdim = 50
 @show _maxlinkdim
 
 # DMRG cutoff
-_cutoff = 1e-8
+_cutoff = 1.0e-8
 
 # Hopping
 t = 1.0
@@ -30,15 +30,15 @@ U = 1.0
 # Make the free fermion Hamiltonian for the up spins
 os_up = OpSum()
 for n in 1:(N - 1)
-  os_up .+= -t, "Cdagup", n, "Cup", n + 1
-  os_up .+= -t, "Cdagup", n + 1, "Cup", n
+    os_up .+= -t, "Cdagup", n, "Cup", n + 1
+    os_up .+= -t, "Cdagup", n + 1, "Cup", n
 end
 
 # Make the free fermion Hamiltonian for the down spins
 os_dn = OpSum()
 for n in 1:(N - 1)
-  os_dn .+= -t, "Cdagdn", n, "Cdn", n + 1
-  os_dn .+= -t, "Cdagdn", n + 1, "Cdn", n
+    os_dn .+= -t, "Cdagdn", n, "Cdn", n + 1
+    os_dn .+= -t, "Cdagdn", n + 1, "Cdn", n
 end
 
 # Hopping Hamiltonians for the up and down spins
@@ -50,17 +50,17 @@ h_dn = hopping_hamiltonian(os_dn)
 Φ_dn = slater_determinant_matrix(h_dn, Nf_dn)
 
 # Create an MPS from the slater determinants.
-s = siteinds("Electron", N; conserve_qns=true)
+s = siteinds("Electron", N; conserve_qns = true)
 println("Making free fermion starting MPS")
 @time ψ0 = slater_determinant_to_mps(
-  s, Φ_up, Φ_dn; eigval_cutoff=1e-4, cutoff=_cutoff, maxdim=_maxlinkdim
+    s, Φ_up, Φ_dn; eigval_cutoff = 1.0e-4, cutoff = _cutoff, maxdim = _maxlinkdim
 )
 @show maxlinkdim(ψ0)
 
 # The total interacting Hamiltonian
 os = os_up + os_dn
 for n in 1:N
-  os .+= U, "Nupdn", n
+    os .+= U, "Nupdn", n
 end
 H = MPO(os, s)
 
@@ -69,7 +69,7 @@ println("Free fermion starting state energy")
 @show inner(ψ0', H, ψ0)
 
 println("\nStart from free fermion state")
-e, ψ = @time dmrg(H, ψ0; nsweeps=5, maxdim=_maxlinkdim, cutoff=_cutoff)
+e, ψ = @time dmrg(H, ψ0; nsweeps = 5, maxdim = _maxlinkdim, cutoff = _cutoff)
 @show e
 @show flux(ψ)
 
@@ -77,6 +77,6 @@ using ITensorGaussianMPS: correlation_matrix_to_gmps, correlation_matrix_to_mps,
 
 Λ_up = correlation_matrix(ψ, "Cdagup", "Cup")
 Λ_dn = correlation_matrix(ψ, "Cdagdn", "Cdn")
-ψ̃0 = correlation_matrix_to_mps(s, Λ_up, Λ_dn; eigval_cutoff=1e-2, maxblocksize=4)
+ψ̃0 = correlation_matrix_to_mps(s, Λ_up, Λ_dn; eigval_cutoff = 1.0e-2, maxblocksize = 4)
 @show inner(ψ̃0, ψ)
 @show inner(ψ̃0', H, ψ̃0)
