+++ test/Project.toml
@@ -1,6 +1,20 @@
 [deps]
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 ITensorGaussianMPS = "2be41995-7c9f-4653-b682-bfa4e7cebb93"
 ITensorMPS = "0d1a4710-d33b-49a5-8f18-73bdf49b47e2"
+ITensorPkgSkeleton = "3d388ab1-018a-49f4-ae50-18094d5f71ea"
 ITensors = "9136182c-28ba-11e9-034c-db9fb085ebd5"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[sources.ITensorGaussianMPS]
+path = ".."
+
+[compat]
+Aqua = "0.8.9"
+ITensorGaussianMPS = "0.1"
+ITensorMPS = "0.3"
+ITensorPkgSkeleton = "0.3.42"
+ITensors = "0.7, 0.8, 0.9"
+LinearAlgebra = "1.10"
+Test = "1.10"
