+++ test/test_basics.jl
@@ -1,0 +1,6 @@
+using ITensorGaussianMPS: ITensorGaussianMPS
+using Test: @test, @testset
+
+@testset "ITensorGaussianMPS" begin
+    # Tests go here.
+end
